Split files for base actions, reducers and selectors.
