+++ ui/src/app/base/actions/index.js
@@ -1,0 +1,3 @@
+export { default as auth } from "./auth";
+export { default as messages } from "./messages";
+export { default as websocket } from "./websocket";
