+++ ui/src/app/base/actions/messages.test.js
@@ -1,22 +1,6 @@
-import { connectWebSocket, messages, fetchAuthUser } from "./actions";
+import messages from "./messages";
 
 describe("base actions", () => {
-  it("should handle connection to a WebSocket", () => {
-    expect(connectWebSocket()).toEqual({
-      type: "WEBSOCKET_CONNECT"
-    });
-  });
-
-  it("should handle fetching the current user", () => {
-    expect(fetchAuthUser()).toEqual({
-      type: "FETCH_AUTH_USER",
-      meta: {
-        model: "user",
-        method: "auth_user"
-      }
-    });
-  });
-
   it("should handle adding a message", () => {
     expect(messages.add("User added", "negative", "Error", true)).toEqual({
       type: "ADD_MESSAGE",
