+++ ui/src/app/base/actions/websocket.js
@@ -1,0 +1,9 @@
+const websocket = {};
+
+websocket.connect = () => {
+  return {
+    type: "WEBSOCKET_CONNECT"
+  };
+};
+
+export default websocket;
