+++ ui/src/app/base/actions/websocket.test.js
@@ -1,0 +1,9 @@
+import websocket from "./websocket";
+
+describe("base actions", () => {
+  it("should handle connection to a WebSocket", () => {
+    expect(websocket.connect()).toEqual({
+      type: "WEBSOCKET_CONNECT"
+    });
+  });
+});
