+++ ui/src/app/base/components/Header/Header.test.js
@@ -1,11 +1,32 @@
-import { shallow } from "enzyme";
+import { MemoryRouter } from "react-router-dom";
+import { mount } from "enzyme";
+import { Provider } from "react-redux";
+import configureStore from "redux-mock-store";
 import React from "react";
 
 import { Header } from "./Header";
 
+const mockStore = configureStore();
+
 describe("Header", () => {
+  let state;
+
   beforeEach(() => {
     process.env.REACT_APP_MAAS_URL = "http://maas.local:5240/MAAS";
+    state = {
+      messages: {
+        items: []
+      },
+      status: {
+        connected: true
+      },
+      user: {
+        auth: {
+          loading: false,
+          user: {}
+        }
+      }
+    };
   });
 
   afterEach(() => {
@@ -13,21 +34,14 @@ describe("Header", () => {
   });
 
   it("renders", () => {
-    const wrapper = shallow(
-      <Header
-        authUser={{
-          email: "test@example.com",
-          first_name: "",
-          global_permissions: ["machine_create"],
-          id: 1,
-          is_superuser: true,
-          last_name: "",
-          sshkeys_count: 0,
-          username: "admin"
-        }}
-        location={{ pathname: "/", hash: "", search: "" }}
-      />
+    const store = mockStore(state);
+    const wrapper = mount(
+      <Provider store={store}>
+        <MemoryRouter initialEntries={[{ pathname: "/" }]}>
+          <Header />
+        </MemoryRouter>
+      </Provider>
     );
-    expect(wrapper).toMatchSnapshot();
+    expect(wrapper.find("Header")).toMatchSnapshot();
   });
 });
