+++ ui/src/app/base/reducers/index.js
@@ -1,0 +1,3 @@
+export { default as auth } from "./auth";
+export { default as messages } from "./messages";
+export { default as status } from "./status";
