+++ ui/src/app/base/selectors/messages.js
@@ -1,0 +1,10 @@
+const messages = {};
+
+/**
+ * Get the global messages.
+ * @param {Object} state - The redux state.
+ * @returns {Array} The list of messages.
+ */
+messages.all = state => state.messages.items;
+
+export default messages;
