+++ ui/src/app/base/selectors/messages.test.js
@@ -1,0 +1,22 @@
+import messages from "./messages";
+
+describe("messages", () => {
+  it("can get all messages", () => {
+    const state = {
+      messages: {
+        items: [
+          {
+            id: 1,
+            message: "User added"
+          }
+        ]
+      }
+    };
+    expect(messages.all(state)).toStrictEqual([
+      {
+        id: 1,
+        message: "User added"
+      }
+    ]);
+  });
+});
