+++ ui/src/app/base/selectors/status.js
@@ -1,0 +1,17 @@
+const status = {};
+
+/**
+ * Whether the websocket is connected.
+ * @param {Object} state - The redux state.
+ * @returns {Boolean} The websocket connected status.
+ */
+status.connected = state => state.status.connected;
+
+/**
+ * Whether there is a websocket error.
+ * @param {Object} state - The redux state.
+ * @returns {Boolean} The websocket error status.
+ */
+status.error = state => state.status.error;
+
+export default status;
