+++ ui/src/app/settings/components/Nav/Nav.test.js
@@ -1,19 +1,24 @@
-import { shallow } from "enzyme";
+import { MemoryRouter } from "react-router-dom";
+import { mount } from "enzyme";
 import React from "react";
 
 import { Nav } from "./Nav";
 
 describe("Nav", () => {
   it("renders", () => {
-    const wrapper = shallow(
-      <Nav location={{ pathname: "/", hash: "", search: "" }} />
+    const wrapper = mount(
+      <MemoryRouter initialEntries={[{ pathname: "/" }]}>
+        <Nav />
+      </MemoryRouter>
     );
-    expect(wrapper).toMatchSnapshot();
+    expect(wrapper.find("Nav")).toMatchSnapshot();
   });
 
   it("can set an active item", () => {
-    const wrapper = shallow(
-      <Nav location={{ pathname: "/users", hash: "", search: "" }} />
+    const wrapper = mount(
+      <MemoryRouter initialEntries={[{ pathname: "/users" }]}>
+        <Nav />
+      </MemoryRouter>
     );
     expect(wrapper.find(".is-active Link").prop("children")).toEqual("Users");
   });
